Got the buttons to change 'base' value
Got the plus and minus buttons to change the networkBaseLevel value so that the buildNetwork function can use it. The displayed value on the webpage is wrong, but I don't think it's because of the plus or minus buttons as they're coded. I set it to display in console, and that seems accurate.